+++ CHANGELOG.md
@@ -10,7 +10,7 @@ this project uses date-based 'snapshot' version identifiers.
 
 ### Added
 
-- Variable `checkpatterns` (issue \#449)
+- Variable `versionpatterns` (issue \#449)
 
 ### Changed